requirements.txt
Flask==3.1.0
msal==1.31.1
azure-storage-blob==12.24.0
azure-identity==1.19.0
python-dotenv==1.0.0